Modul2/uwierzytelnianie_i_autoryzacja/Lekcja 1/requirements.txt
fastapi
uvicorn[standard]
fastapi-jwt-auth
python-jose
sqlalchemy
passlib 
python-multipart
bcrypt==4.0.1